Look at all those chords! Imagine what Status Quo could have become if they'd known them!

+++ Assets/Music.cs
@@ -77,7 +77,7 @@ public class Music : MonoBehaviour
 
     void CalcMeasureLength()
     {
-        MeasureLength = 60f/BeatsPerMinute * (float)BeatsPerMeasure;
+        MeasureLength = 60f/BeatsPerMinute * BeatsPerMeasure;
     }
 
     // Use this for initialization
